Align DSv4 MegaMoE modeling structure
