+++ src/aiconfigurator/sdk/operations.py
@@ -2,6 +2,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import logging
+from collections.abc import Iterable, Sequence
+from dataclasses import dataclass
 from typing import Optional
 
 from aiconfigurator.sdk import common
@@ -9,6 +11,366 @@ from aiconfigurator.sdk.perf_database import PerfDatabase
 from aiconfigurator.sdk.performance_result import PerformanceResult
 
 logger = logging.getLogger(__name__)
+
+
+@dataclass(frozen=True)
+class MegaMoETrafficEstimate:
+    route_matrix: tuple[tuple[int, ...], ...]
+    hidden_size: int
+    total_remote_edges: int
+    owner_remote_edges: tuple[int, ...]
+    endpoint_remote_edges: tuple[int, ...]
+    bottleneck_owner_rank: int
+    bottleneck_endpoint_rank: int
+    primary_bytes_per_remote_selection: int
+    total_primary_bytes: int
+    owner_primary_bytes: tuple[int, ...]
+    endpoint_primary_bytes: tuple[int, ...]
+    bottleneck_primary_bytes: int
+    bottleneck_endpoint_primary_bytes: int
+    metadata_bytes_per_remote_selection: int
+    total_metadata_bytes: int
+    owner_metadata_bytes: tuple[int, ...]
+    endpoint_metadata_bytes: tuple[int, ...]
+    bottleneck_metadata_bytes: int
+    bottleneck_endpoint_metadata_bytes: int
+
+
+@dataclass(frozen=True)
+class MegaMoECommunicationEstimate:
+    """Communication term for the fused DeepGEMM MegaMoE kernel.
+
+    ``overlappable_ms`` is the hot-path communication term that belongs under
+    ``max(core_compute_ms, overlappable_ms)``. ``tail_ms`` is reported
+    separately for callers that model a cleanup/sync tail outside that max.
+    """
+
+    traffic: MegaMoETrafficEstimate
+    effective_nvlink_bandwidth_bps: float
+    nvl_barrier_latency_us: float
+    nvl_barrier_count: int
+    cleanup_barrier_count: int
+    include_metadata: bool
+    bottleneck_rank: int
+    data_bytes: int
+    data_ms: float
+    fixed_overlappable_latency_ms: float
+    barrier_ms: float
+    cleanup_barrier_ms: float
+    overlappable_ms: float
+    tail_ms: float
+    unoverlapped_ms: float
+    total_ms: float
+
+
+def owner_rank_for_expert(expert_id: int, *, num_experts: int, moe_ep_size: int) -> int:
+    if moe_ep_size <= 0:
+        raise ValueError("moe_ep_size must be positive")
+    if num_experts <= 0:
+        raise ValueError("num_experts must be positive")
+    if num_experts % moe_ep_size != 0:
+        raise ValueError("num_experts must be divisible by moe_ep_size")
+    if expert_id < 0 or expert_id >= num_experts:
+        raise ValueError(f"expert_id out of range: {expert_id}")
+
+    experts_per_rank = num_experts // moe_ep_size
+    return expert_id // experts_per_rank
+
+
+def build_route_matrix(
+    topk_ids_by_src_rank: Sequence[Sequence[Sequence[int]]],
+    *,
+    num_experts: int,
+    moe_ep_size: int,
+) -> tuple[tuple[int, ...], ...]:
+    """Build routes[src_rank][dst_rank] from per-source-rank top-k IDs."""
+    if len(topk_ids_by_src_rank) != moe_ep_size:
+        raise ValueError("topk_ids_by_src_rank length must equal moe_ep_size")
+
+    routes = [[0 for _ in range(moe_ep_size)] for _ in range(moe_ep_size)]
+    for src_rank, rank_topk in enumerate(topk_ids_by_src_rank):
+        for token_topk in rank_topk:
+            for expert_id in token_topk:
+                expert_id = int(expert_id)
+                if expert_id == -1:
+                    continue
+                dst_rank = owner_rank_for_expert(
+                    expert_id,
+                    num_experts=num_experts,
+                    moe_ep_size=moe_ep_size,
+                )
+                routes[src_rank][dst_rank] += 1
+    return tuple(tuple(row) for row in routes)
+
+
+def route_matrix_from_flat_assignments(
+    assignments: Iterable[tuple[int, int]],
+    *,
+    num_experts: int,
+    moe_ep_size: int,
+) -> tuple[tuple[int, ...], ...]:
+    """Build a route matrix from ``(src_rank, expert_id)`` assignments."""
+    if moe_ep_size <= 0:
+        raise ValueError("moe_ep_size must be positive")
+    routes = [[0 for _ in range(moe_ep_size)] for _ in range(moe_ep_size)]
+    for src_rank, expert_id in assignments:
+        src_rank = int(src_rank)
+        if src_rank < 0 or src_rank >= moe_ep_size:
+            raise ValueError(f"src_rank out of range: {src_rank}")
+        expert_id = int(expert_id)
+        if expert_id == -1:
+            continue
+        dst_rank = owner_rank_for_expert(
+            expert_id,
+            num_experts=num_experts,
+            moe_ep_size=moe_ep_size,
+        )
+        routes[src_rank][dst_rank] += 1
+    return tuple(tuple(row) for row in routes)
+
+
+def normalize_route_matrix(route_matrix: Sequence[Sequence[int]]) -> tuple[tuple[int, ...], ...]:
+    if not route_matrix:
+        raise ValueError("route_matrix must be non-empty")
+
+    width = len(route_matrix)
+    normalized: list[tuple[int, ...]] = []
+    for row in route_matrix:
+        if len(row) != width:
+            raise ValueError("route_matrix must be square")
+        normalized_row = tuple(int(value) for value in row)
+        if any(value < 0 for value in normalized_row):
+            raise ValueError("route_matrix entries must be non-negative")
+        normalized.append(normalized_row)
+    return tuple(normalized)
+
+
+def estimate_megamoe_traffic(
+    route_matrix: Sequence[Sequence[int]],
+    *,
+    hidden_size: int,
+    quant_group_size: int = 32,
+) -> MegaMoETrafficEstimate:
+    """Estimate DSv4 MegaMoE inter-rank traffic from a route matrix.
+
+    Primary bytes follow DeepGEMM's benchmark convention: remote dispatch pulls
+    FP8 activation (``hidden`` bytes) and combine writes BF16 output
+    (``2 * hidden`` bytes), so one remote token/expert selection costs
+    ``3 * hidden`` primary bytes.
+    """
+    if hidden_size <= 0:
+        raise ValueError("hidden_size must be positive")
+    if quant_group_size <= 0:
+        raise ValueError("quant_group_size must be positive")
+
+    routes = normalize_route_matrix(route_matrix)
+    ep_size = len(routes)
+    primary_per_edge = 3 * hidden_size
+    metadata_per_edge = (hidden_size + quant_group_size - 1) // quant_group_size + 8
+
+    owner_edges = [0 for _ in range(ep_size)]
+    endpoint_edges = [0 for _ in range(ep_size)]
+    total_remote_edges = 0
+    for src_rank, row in enumerate(routes):
+        for dst_rank, edges in enumerate(row):
+            if src_rank == dst_rank:
+                continue
+            total_remote_edges += edges
+            owner_edges[dst_rank] += edges
+            endpoint_edges[src_rank] += edges
+
+    owner_primary = tuple(edges * primary_per_edge for edges in owner_edges)
+    endpoint_primary = tuple(edges * primary_per_edge for edges in endpoint_edges)
+    owner_metadata = tuple(edges * metadata_per_edge for edges in owner_edges)
+    endpoint_metadata = tuple(edges * metadata_per_edge for edges in endpoint_edges)
+    bottleneck_owner_rank = max(range(ep_size), key=lambda rank: owner_edges[rank])
+    bottleneck_endpoint_rank = max(range(ep_size), key=lambda rank: endpoint_edges[rank])
+
+    return MegaMoETrafficEstimate(
+        route_matrix=routes,
+        hidden_size=hidden_size,
+        total_remote_edges=total_remote_edges,
+        owner_remote_edges=tuple(owner_edges),
+        endpoint_remote_edges=tuple(endpoint_edges),
+        bottleneck_owner_rank=bottleneck_owner_rank,
+        bottleneck_endpoint_rank=bottleneck_endpoint_rank,
+        primary_bytes_per_remote_selection=primary_per_edge,
+        total_primary_bytes=total_remote_edges * primary_per_edge,
+        owner_primary_bytes=owner_primary,
+        endpoint_primary_bytes=endpoint_primary,
+        bottleneck_primary_bytes=max(owner_primary, default=0),
+        bottleneck_endpoint_primary_bytes=max(endpoint_primary, default=0),
+        metadata_bytes_per_remote_selection=metadata_per_edge,
+        total_metadata_bytes=total_remote_edges * metadata_per_edge,
+        owner_metadata_bytes=owner_metadata,
+        endpoint_metadata_bytes=endpoint_metadata,
+        bottleneck_metadata_bytes=max(owner_metadata, default=0),
+        bottleneck_endpoint_metadata_bytes=max(endpoint_metadata, default=0),
+    )
+
+
+def estimate_megamoe_communication_ms(
+    route_matrix: Sequence[Sequence[int]],
+    *,
+    hidden_size: int,
+    effective_nvlink_bandwidth_bps: float,
+    nvl_barrier_latency_us: float,
+    nvl_barrier_count: int = 2,
+    cleanup_barrier_count: int = 1,
+    fixed_overlappable_latency_ms: float = 0.0,
+    include_metadata: bool = False,
+    quant_group_size: int = 32,
+) -> MegaMoECommunicationEstimate:
+    if effective_nvlink_bandwidth_bps <= 0:
+        raise ValueError("effective_nvlink_bandwidth_bps must be positive")
+    if nvl_barrier_latency_us < 0:
+        raise ValueError("nvl_barrier_latency_us must be non-negative")
+    if nvl_barrier_count < 0:
+        raise ValueError("nvl_barrier_count must be non-negative")
+    if cleanup_barrier_count < 0:
+        raise ValueError("cleanup_barrier_count must be non-negative")
+    if fixed_overlappable_latency_ms < 0:
+        raise ValueError("fixed_overlappable_latency_ms must be non-negative")
+
+    traffic = estimate_megamoe_traffic(
+        route_matrix,
+        hidden_size=hidden_size,
+        quant_group_size=quant_group_size,
+    )
+    data_bytes = traffic.bottleneck_primary_bytes
+    if include_metadata:
+        data_bytes += traffic.bottleneck_metadata_bytes
+
+    data_ms = data_bytes / effective_nvlink_bandwidth_bps * 1000.0
+    barrier_ms = nvl_barrier_count * nvl_barrier_latency_us / 1000.0
+    cleanup_barrier_ms = cleanup_barrier_count * nvl_barrier_latency_us / 1000.0
+    overlappable_ms = data_ms + barrier_ms + fixed_overlappable_latency_ms
+    tail_ms = cleanup_barrier_ms
+    unoverlapped_ms = overlappable_ms + tail_ms
+    return MegaMoECommunicationEstimate(
+        traffic=traffic,
+        effective_nvlink_bandwidth_bps=float(effective_nvlink_bandwidth_bps),
+        nvl_barrier_latency_us=float(nvl_barrier_latency_us),
+        nvl_barrier_count=int(nvl_barrier_count),
+        cleanup_barrier_count=int(cleanup_barrier_count),
+        include_metadata=include_metadata,
+        bottleneck_rank=traffic.bottleneck_owner_rank,
+        data_bytes=data_bytes,
+        data_ms=data_ms,
+        fixed_overlappable_latency_ms=float(fixed_overlappable_latency_ms),
+        barrier_ms=barrier_ms,
+        cleanup_barrier_ms=cleanup_barrier_ms,
+        overlappable_ms=overlappable_ms,
+        tail_ms=tail_ms,
+        unoverlapped_ms=unoverlapped_ms,
+        total_ms=unoverlapped_ms,
+    )
+
+
+def estimate_megamoe_communication_from_effective_bandwidth_model_ms(
+    route_matrix: Sequence[Sequence[int]],
+    *,
+    hidden_size: int,
+    num_tokens_per_rank: int,
+    bandwidth_model,
+    nvl_barrier_latency_us: float = 0.0,
+    nvl_barrier_count: int = 0,
+    cleanup_barrier_count: int = 0,
+    include_metadata: bool = False,
+    quant_group_size: int = 32,
+) -> MegaMoECommunicationEstimate:
+    """Estimate communication from a token-dependent effective BW model."""
+    routes = normalize_route_matrix(route_matrix)
+    if len(routes) != bandwidth_model.ep_size:
+        expected_shape = f"{bandwidth_model.ep_size}x{bandwidth_model.ep_size}"
+        raise ValueError(f"bandwidth model {bandwidth_model.name!r} requires a {expected_shape} route matrix")
+    if num_tokens_per_rank <= 0:
+        raise ValueError("num_tokens_per_rank must be positive")
+
+    return estimate_megamoe_communication_ms(
+        routes,
+        hidden_size=hidden_size,
+        effective_nvlink_bandwidth_bps=bandwidth_model.calibrated_bandwidth_bps(num_tokens_per_rank),
+        nvl_barrier_latency_us=nvl_barrier_latency_us,
+        nvl_barrier_count=nvl_barrier_count,
+        cleanup_barrier_count=cleanup_barrier_count,
+        fixed_overlappable_latency_ms=bandwidth_model.fixed_overlappable_latency_ms,
+        include_metadata=include_metadata,
+        quant_group_size=quant_group_size,
+    )
+
+
+def estimate_megamoe_communication_from_perf_database_ms(
+    route_matrix: Sequence[Sequence[int]],
+    *,
+    database: PerfDatabase,
+    hidden_size: int,
+    inter_size: int,
+    topk: int,
+    num_experts: int,
+    moe_ep_size: int,
+    routing_mode: str,
+    power_law_alpha: float | None,
+    num_tokens_per_rank: int,
+    kernel_source: str = "DeepGEMM_fp8_fp4_mega_moe",
+    nvl_barrier_latency_us: float = 0.0,
+    nvl_barrier_count: int = 0,
+    cleanup_barrier_count: int = 0,
+    include_metadata: bool = False,
+    quant_group_size: int = 32,
+) -> MegaMoECommunicationEstimate:
+    """Estimate communication using an effective-BW model loaded by PerfDatabase."""
+    bandwidth_model = database.query_dsv4_megamoe_effective_bandwidth_model(
+        hidden_size=hidden_size,
+        inter_size=inter_size,
+        topk=topk,
+        num_experts=num_experts,
+        moe_ep_size=moe_ep_size,
+        routing_mode=routing_mode,
+        power_law_alpha=power_law_alpha,
+        kernel_source=kernel_source,
+    )
+    return estimate_megamoe_communication_from_effective_bandwidth_model_ms(
+        route_matrix,
+        hidden_size=hidden_size,
+        num_tokens_per_rank=num_tokens_per_rank,
+        bandwidth_model=bandwidth_model,
+        nvl_barrier_latency_us=nvl_barrier_latency_us,
+        nvl_barrier_count=nvl_barrier_count,
+        cleanup_barrier_count=cleanup_barrier_count,
+        include_metadata=include_metadata,
+        quant_group_size=quant_group_size,
+    )
+
+
+def compose_megamoe_routed_latency_ms(
+    *,
+    local_routing_prep_ms: float,
+    core_compute_ms: float,
+    comm_ms: Optional[float] = None,
+    sync_tail_ms: float = 0.0,
+    comm_estimate: Optional[MegaMoECommunicationEstimate] = None,
+) -> float:
+    """Compose routed MegaMoE latency with communication-compute overlap."""
+    if (comm_ms is None) == (comm_estimate is None):
+        raise ValueError("exactly one of comm_ms or comm_estimate must be provided")
+    if comm_estimate is not None:
+        comm_ms = comm_estimate.overlappable_ms
+        sync_tail_ms += comm_estimate.tail_ms
+
+    values = {
+        "local_routing_prep_ms": local_routing_prep_ms,
+        "core_compute_ms": core_compute_ms,
+        "comm_ms": comm_ms,
+        "sync_tail_ms": sync_tail_ms,
+    }
+    for name, value in values.items():
+        if value is None:
+            raise ValueError(f"{name} must be provided")
+        if value < 0:
+            raise ValueError(f"{name} must be non-negative")
+
+    return local_routing_prep_ms + max(core_compute_ms, comm_ms) + sync_tail_ms
 
 
 class Operation:
@@ -35,6 +397,88 @@ class Operation:
 
     def get_weights(self, **kwargs):
         raise NotImplementedError
+
+
+class Dsv4MegaMoEDispatch(Operation):
+    """DSv4 MegaMoE fused-kernel communication term.
+
+    This follows the existing MoE split in AIC: the operation owns the
+    communication-side query and delegates measured calibration lookup to
+    ``PerfDatabase``.  The route matrix is provided by the model/collector
+    layer because it depends on the target routing distribution.
+    """
+
+    def __init__(
+        self,
+        name: str,
+        scale_factor: float,
+        hidden_size: int,
+        inter_size: int,
+        topk: int,
+        num_experts: int,
+        moe_ep_size: int,
+        routing_mode: str,
+        power_law_alpha: Optional[float],
+        kernel_source: str = "DeepGEMM_fp8_fp4_mega_moe",
+        nvl_barrier_latency_us: float = 0.0,
+        nvl_barrier_count: int = 0,
+        cleanup_barrier_count: int = 0,
+        include_metadata: bool = False,
+        quant_group_size: int = 32,
+        include_tail: bool = False,
+        **kwargs,
+    ) -> None:
+        super().__init__(name, scale_factor)
+        self._hidden_size = hidden_size
+        self._inter_size = inter_size
+        self._topk = topk
+        self._num_experts = num_experts
+        self._moe_ep_size = moe_ep_size
+        self._routing_mode = routing_mode
+        self._power_law_alpha = power_law_alpha
+        self._kernel_source = kernel_source
+        self._nvl_barrier_latency_us = nvl_barrier_latency_us
+        self._nvl_barrier_count = nvl_barrier_count
+        self._cleanup_barrier_count = cleanup_barrier_count
+        self._include_metadata = include_metadata
+        self._quant_group_size = quant_group_size
+        self._include_tail = include_tail
+        self._weights = 0.0
+
+    def query(self, database: PerfDatabase, **kwargs) -> PerformanceResult:
+        route_matrix = kwargs.get("route_matrix")
+        if route_matrix is None:
+            raise ValueError("Dsv4MegaMoEDispatch requires route_matrix in query kwargs")
+
+        num_tokens_per_rank = kwargs.get("num_tokens_per_rank")
+        if num_tokens_per_rank is None:
+            num_tokens_per_rank = kwargs.get("x")
+        if num_tokens_per_rank is None:
+            raise ValueError("Dsv4MegaMoEDispatch requires num_tokens_per_rank or x in query kwargs")
+
+        comm = estimate_megamoe_communication_from_perf_database_ms(
+            route_matrix,
+            database=database,
+            hidden_size=self._hidden_size,
+            inter_size=self._inter_size,
+            topk=self._topk,
+            num_experts=self._num_experts,
+            moe_ep_size=self._moe_ep_size,
+            routing_mode=self._routing_mode,
+            power_law_alpha=self._power_law_alpha,
+            num_tokens_per_rank=int(num_tokens_per_rank),
+            kernel_source=self._kernel_source,
+            nvl_barrier_latency_us=self._nvl_barrier_latency_us,
+            nvl_barrier_count=self._nvl_barrier_count,
+            cleanup_barrier_count=self._cleanup_barrier_count,
+            include_metadata=self._include_metadata,
+            quant_group_size=self._quant_group_size,
+        )
+        latency = comm.overlappable_ms + (comm.tail_ms if self._include_tail else 0.0)
+        return PerformanceResult(latency * self._scale_factor, energy=0.0)
+
+    def get_weights(self, **kwargs):
+        return self._weights * self._scale_factor
 
 
 class CustomAllReduce(Operation):
